pyproject.toml
[tool.poetry]
name = "translation-service-example"
version = "0.1.0"
description = ""
authors = ["Andrew Hoyt <ahoyt41@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
grpcio-tools = "^1.62.1"
torch = {version = "^2.2.2+cpu", source = "torch-cpu"}
torchvision = {version = "^0.17.2+cpu", source = "torch-cpu"}
torchaudio = {version = "^2.2.2+cpu", source = "torch-cpu"}
ctranslate2 = "4.1.0"
googleapis-common-protos = "^1.63.0"
sentencepiece = "^0.2.0"
pydantic = "^2.7.0"


[[tool.poetry.source]]
name = "torch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"


[tool.poetry.group.opennmt.dependencies]
opennmt-py = "==2.*"


[tool.poetry.group.grpc.dependencies]
grpcio-tools = "^1.62.1"
mypy-protobuf = "^3.6.0"

[tool.poetry.scripts]
server = "translation_service_example.server:main"
client = "translation_service_example.client:main"

[tool.poetry.group.dev.dependencies]
grpc-stubs = "^1.53.0.5"
mypy = "^1.9.0"
ruff = "^0.3.5"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
